+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/facet/impl/AppEngineStandardCommunityWebIntegration.java
@@ -16,10 +16,10 @@
 
 package com.google.cloud.tools.intellij.appengine.facet.impl;
 
-import com.google.cloud.tools.intellij.appengine.facet.AppEngineWebIntegration;
+import com.google.cloud.tools.intellij.appengine.facet.AppEngineStandardWebIntegration;
 
-import com.intellij.execution.configurations.ModuleRunConfiguration;
 import com.intellij.framework.addSupport.FrameworkSupportInModuleProvider;
+import com.intellij.ide.util.frameworkSupport.FrameworkRole;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.project.Project;
@@ -45,9 +45,10 @@ import java.util.List;
 /**
  * @author nik.
  */
-public class AppEngineCommunityWebIntegration extends AppEngineWebIntegration {
+public class AppEngineStandardCommunityWebIntegration extends AppEngineStandardWebIntegration {
 
-  private static final Logger LOG = Logger.getInstance(AppEngineCommunityWebIntegration.class);
+  private static final Logger LOG
+      = Logger.getInstance(AppEngineStandardCommunityWebIntegration.class);
 
   @Nullable
   @Override
@@ -77,13 +78,20 @@ public class AppEngineCommunityWebIntegration extends AppEngineWebIntegration {
     return null;
   }
 
+  @Nullable
   @Override
-  public void setupJpaSupport(@NotNull Module module, @NotNull VirtualFile persistenceXml) {
+  public String getUnderlyingFrameworkTypeId() {
+    return null;
+  }
+
+  @NotNull
+  @Override
+  public FrameworkRole[] getFrameworkRoles() {
+    return new FrameworkRole[0];
   }
 
   @Override
-  public void setupRunConfiguration(@Nullable Artifact artifact, @NotNull Project project,
-      ModuleRunConfiguration existingConfiguration) {
+  public void setupJpaSupport(@NotNull Module module, @NotNull VirtualFile persistenceXml) {
   }
 
   @Override