+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/sdk/CloudSdkPanel.java
@@ -32,6 +32,7 @@ import com.intellij.ui.JBColor;
 
 import org.jetbrains.annotations.NotNull;
 
+import java.util.Objects;
 import java.util.Set;
 import java.util.TreeSet;
 
@@ -74,8 +75,6 @@ public class CloudSdkPanel {
             checkSdkInBackground();
           }
         });
-
-    checkSdkInBackground();
   }
 
   private void checkSdkInBackground() {
@@ -85,34 +84,42 @@ public class CloudSdkPanel {
 
   @VisibleForTesting
   protected void checkSdk(String path) {
+    String message = buildSdkMessage(path, true /*htmlEnabled*/);
+
+    if (!StringUtil.isEmpty(message)) {
+      showWarning(message);
+    } else {
+      hideWarning();
+    }
+  }
+
+  public String buildSdkMessage(String path, boolean htmlEnabled) {
     if (StringUtil.isEmpty(path)) {
-      String warningText = GctBundle.message("appengine.cloudsdk.location.missing.message")
-          + " " + getCloudSdkDownloadMessage();
-      showWarning(warningText , false /* setSdkDirectoryErrorState */);
-      return;
+      String missingMessage = GctBundle.message("appengine.cloudsdk.location.missing.message");
+
+      return htmlEnabled
+          ? missingMessage + " " + getCloudSdkDownloadMessage()
+          : missingMessage;
     }
 
     CloudSdkService sdkService = CloudSdkService.getInstance();
+    // Use a sorted set to guarantee consistent ordering of CloudSdkValidationResults.
+    Set<CloudSdkValidationResult> validationResults =
+        new TreeSet<>(sdkService.validateCloudSdk(path));
 
-    if (sdkService.isValidCloudSdk(path)) {
-      hideWarning();
-    } else {
-      // Use a sorted set to guarantee consistent ordering of CloudSdkValidationResults.
-      Set<CloudSdkValidationResult> validationResults =
-          new TreeSet<>(sdkService.validateCloudSdk(path));
-
+    if (!validationResults.isEmpty()) {
       // Display all validation results as a list.
       StringBuilder builder = new StringBuilder();
-      boolean containsErrors = false;
 
       boolean isFirst = true;
       for (CloudSdkValidationResult validationResult : validationResults) {
-        containsErrors |= validationResult.isError();
 
         String message;
         if (validationResult == CloudSdkValidationResult.CLOUD_SDK_NOT_FOUND) {
           // If the cloud sdk is not found, provide a download URL.
-          message = validationResult.getMessage() + " " + getCloudSdkDownloadMessage();
+          message = htmlEnabled
+              ? validationResult.getMessage() + " " + getCloudSdkDownloadMessage()
+              : validationResult.getMessage();
         } else {
           // Otherwise, just use the existing message.
           message = validationResult.getMessage();
@@ -126,22 +133,21 @@ public class CloudSdkPanel {
         isFirst = false;
       }
 
-      showWarning(builder.toString(), containsErrors);
+      return builder.toString();
     }
+
+    return null;
   }
 
   @VisibleForTesting
-  protected void showWarning(final String message, final boolean setSdkDirectoryErrorState) {
+  protected void showWarning(final String message) {
     invokePanelValidationUpdate(new Runnable() {
       @Override
       public void run() {
         warningMessage.setText(message);
         warningMessage.setVisible(true);
-        warningMessage.updateUI();
         warningIcon.setVisible(true);
-        if (setSdkDirectoryErrorState) {
-          cloudSdkDirectoryField.getTextField().setForeground(JBColor.red);
-        }
+        cloudSdkDirectoryField.getTextField().setForeground(JBColor.red);
       }
     });
   }
@@ -153,16 +159,13 @@ public class CloudSdkPanel {
       public void run() {
         cloudSdkDirectoryField.getTextField().setForeground(JBColor.black);
         warningIcon.setVisible(false);
-        warningIcon.setVisible(false);
         warningMessage.setVisible(false);
-        cloudSdkPanel.updateUI();
       }
     });
   }
 
   private void invokePanelValidationUpdate(Runnable runnable) {
-    ApplicationManager.getApplication().invokeLater(runnable,
-        ModalityState.stateForComponent(cloudSdkPanel));
+    ApplicationManager.getApplication().invokeLater(runnable, ModalityState.any());
   }
 
   @VisibleForTesting
@@ -186,7 +189,15 @@ public class CloudSdkPanel {
   }
 
   public void apply() throws ConfigurationException {
-    CloudSdkService.getInstance().setSdkHomePath(getCloudSdkDirectoryText());
+    CloudSdkService sdkService = CloudSdkService.getInstance();
+
+    if (sdkService.validateCloudSdk(getCloudSdkDirectoryText())
+        .contains(CloudSdkValidationResult.MALFORMED_PATH)) {
+      throw new ConfigurationException(
+          GctBundle.message("appengine.cloudsdk.location.badchars.message"));
+    }
+
+    sdkService.setSdkHomePath(getCloudSdkDirectoryText());
   }
 
   public void reset() {
@@ -200,8 +211,14 @@ public class CloudSdkPanel {
     return cloudSdkDirectoryField.getText();
   }
 
+  /**
+   * Sets the Cloud SDK directory text. Performs no action if the current value of the Cloud SDK
+   * directory field is already equal to the path.
+   */
   public void setCloudSdkDirectoryText(String path) {
-    cloudSdkDirectoryField.setText(path);
+    if (!Objects.equals(cloudSdkDirectoryField.getText(), path)) {
+      cloudSdkDirectoryField.setText(path);
+    }
   }
 
   @NotNull