+++ google-cloud-tools-plugin/testSrc/com/google/cloud/tools/intellij/appengine/sdk/CloudSdkPanelTest.java
@@ -24,6 +24,8 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import com.google.cloud.tools.intellij.util.GctBundle;
+
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.testFramework.PlatformTestCase;
 import com.intellij.util.containers.HashSet;
@@ -53,11 +55,8 @@ public class CloudSdkPanelTest extends PlatformTestCase {
   private static final String MISSING_SDK_DIR_WARNING =
       "Cloud SDK home directory is not specified. "
           + CLOUD_SDK_DOWNLOAD_LINK;
-  private static final String INVALID_SDK_DIR_WARNING = "No Cloud SDK was found in this directory. "
+  private static final String INVALID_SDK_DIR_WARNING = "No Cloud SDK was found in the specified directory. "
       + CLOUD_SDK_DOWNLOAD_LINK;
-  private static final String UNSUPPORTED_SDK_WARNING = "The configured Google Cloud SDK is out of "
-      + "date. Version 131.0.0 is the minimum recommended version for use with the Google Cloud Tools Plugin. To update, "
-      + "run \"gcloud components update\".";
 
   @Override
   public void setUp() throws Exception {
@@ -80,14 +79,14 @@ public class CloudSdkPanelTest extends PlatformTestCase {
   public void testCheckSdk_nullSdk() throws InterruptedException {
     when(cloudSdkService.isValidCloudSdk(null)).thenReturn(false);
     panel.checkSdk(null);
-    verify(panel, times(1)).showWarning(eq(MISSING_SDK_DIR_WARNING), eq(false));
+    verify(panel, times(1)).showWarning(eq(MISSING_SDK_DIR_WARNING));
     verify(panel, times(0)).hideWarning();
   }
   @Test
   public void testCheckSdk_emptySdk() throws InterruptedException {
     when(cloudSdkService.isValidCloudSdk("")).thenReturn(false);
     panel.checkSdk("");
-    verify(panel, times(1)).showWarning(eq(MISSING_SDK_DIR_WARNING), eq(false));
+    verify(panel, times(1)).showWarning(eq(MISSING_SDK_DIR_WARNING));
     verify(panel, times(0)).hideWarning();
   }
 
@@ -96,7 +95,7 @@ public class CloudSdkPanelTest extends PlatformTestCase {
     setValidateCloudSdkResponse(CloudSdkValidationResult.CLOUD_SDK_NOT_FOUND);
     when(cloudSdkService.isValidCloudSdk("/non/empty/path")).thenReturn(false);
     panel.checkSdk("/non/empty/path");
-    verify(panel, times(1)).showWarning(eq(INVALID_SDK_DIR_WARNING), eq(true));
+    verify(panel, times(1)).showWarning(eq(INVALID_SDK_DIR_WARNING));
     verify(panel, times(0)).hideWarning();
   }
 
@@ -105,7 +104,8 @@ public class CloudSdkPanelTest extends PlatformTestCase {
     setValidateCloudSdkResponse(CloudSdkValidationResult.CLOUD_SDK_VERSION_NOT_SUPPORTED);
     when(cloudSdkService.isValidCloudSdk("/non/empty/path")).thenReturn(false);
     panel.checkSdk("/non/empty/path");
-    verify(panel, times(1)).showWarning(eq(UNSUPPORTED_SDK_WARNING), eq(false));
+    verify(panel, times(1)).showWarning(
+        eq(CloudSdkValidationResult.CLOUD_SDK_VERSION_NOT_SUPPORTED.getMessage()));
     verify(panel, times(0)).hideWarning();
   }
 
@@ -115,10 +115,11 @@ public class CloudSdkPanelTest extends PlatformTestCase {
         CloudSdkValidationResult.CLOUD_SDK_NOT_FOUND);
     when(cloudSdkService.isValidCloudSdk("/non/empty/path")).thenReturn(false);
 
-    String expectedMessage = INVALID_SDK_DIR_WARNING + "<p>" + UNSUPPORTED_SDK_WARNING + "</p>";
+    String expectedMessage = INVALID_SDK_DIR_WARNING + "<p>" +
+        CloudSdkValidationResult.CLOUD_SDK_VERSION_NOT_SUPPORTED.getMessage() + "</p>";
 
     panel.checkSdk("/non/empty/path");
-    verify(panel, times(1)).showWarning(eq(expectedMessage), eq(true));
+    verify(panel, times(1)).showWarning(eq(expectedMessage));
     verify(panel, times(0)).hideWarning();
   }
 
@@ -127,7 +128,7 @@ public class CloudSdkPanelTest extends PlatformTestCase {
     when(cloudSdkService.isValidCloudSdk("/non/empty/path")).thenReturn(true);
     setValidateCloudSdkResponse();
     panel.checkSdk("/non/empty/path");
-    verify(panel, times(0)).showWarning(any(String.class), anyBoolean());
+    verify(panel, times(0)).showWarning(any(String.class));
     verify(panel, times(1)).hideWarning();
   }